JAMES-1885: Move TestId from mailbox-store to mailbox-api and its refactor
