+++ mailbox/api/src/test/java/org/apache/james/mailbox/model/TestId.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.mailbox.store;
+package org.apache.james.mailbox.model;
 
 import org.apache.james.mailbox.model.MailboxId;
