+++ mailbox/store/src/test/java/org/apache/james/mailbox/store/TestIdDeserializer.java
@@ -19,6 +19,7 @@
 
 package org.apache.james.mailbox.store;
 
+import org.apache.james.mailbox.model.TestId;
 import org.apache.james.mailbox.store.mail.model.MailboxIdDeserializer;
 
 public class TestIdDeserializer implements MailboxIdDeserializer {
